requirements.txt
Flask~=3.0.3
Authlib==1.3.2
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.9.0
requests==2.32.3
Flask-Cors==5.0.0
Werkzeug~=3.0.4
config~=0.5.1
mongomock~=4.1.2